python/requirements.txt
# requirements.txt
selenium==4.15.0
undetected-chromedriver==3.5.4
selenium-stealth==1.0.6
requests==2.31.0
beautifulsoup4==4.12.2
fake-useragent==1.4.0
python-dotenv==1.0.0
webdriver-manager==4.0.1
pyautogui==0.9.54
pandas==2.1.3
openpyxl==3.1.2
pillow==10.1.0
pyvirtualdisplay==3.0